Move PageWrapper out of SwapLayout
