+++ src/pages/About.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { ContentPage } from 'components/layout/SwapLayout/PageWrapper'
+import { ContentPage } from 'components/layout'
 import addressesByNetwork from 'api/deposit/batchExchangeAddresses'
 import { getNetworkFromId } from '@gnosis.pm/dex-js'
 import { EtherscanLink } from 'components/common/EtherscanLink'
