+++ src/pages/Strategies.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import PoolingWidget from 'components/PoolingWidget'
-import { PageWrapper } from 'components/layout/SwapLayout/PageWrapper'
+import { PageWrapper } from 'components/layout/PageWrapper'
 
 const Strategies: React.FC = () => (
   <PageWrapper>
